+++ CHANGELOG.md
@@ -22,3 +22,4 @@ and this project adheres to [Semantic Versioning v2.0.0](https://semver.org/spec
 - Update key auth vm module
 - Update pass auth vm module
 - Add Custom image key auth module
+- Add Custom image pass auth module